[#709] Organize CLI commands into separate files
