+++ iceoryx2-cli/iox2-node/src/command/details.rs
@@ -1,4 +1,4 @@
-// Copyright (c) 2024 Contributors to the Eclipse Foundation
+// Copyright (c) 2025 Contributors to the Eclipse Foundation
 //
 // See the NOTICE file(s) distributed with this work for additional
 // information regarding copyright ownership.
@@ -15,34 +15,15 @@ use iceoryx2::prelude::*;
 use iceoryx2_cli::filter::Filter;
 use iceoryx2_cli::filter::NodeIdentifier;
 use iceoryx2_cli::output::NodeDescription;
-use iceoryx2_cli::output::NodeDescriptor;
-use iceoryx2_cli::output::NodeList;
 use iceoryx2_cli::Format;
 
 use crate::cli::OutputFilter;
 
-pub fn list(filter: OutputFilter, format: Format) -> Result<()> {
-    let mut nodes = Vec::<NodeDescriptor>::new();
-    Node::<ipc::Service>::list(Config::global_config(), |node| {
-        if filter.matches(&node) {
-            nodes.push(NodeDescriptor::from(&node));
-        }
-        CallbackProgression::Continue
-    })
-    .context("failed to retrieve nodes")?;
-
-    println!(
-        "{}",
-        format.as_string(&NodeList {
-            num: nodes.len(),
-            details: nodes
-        })?
-    );
-
-    Ok(())
-}
-
-pub fn details(identifier: NodeIdentifier, filter: OutputFilter, format: Format) -> Result<()> {
+pub(crate) fn details(
+    identifier: NodeIdentifier,
+    filter: OutputFilter,
+    format: Format,
+) -> Result<()> {
     let mut error: Option<Error> = None;
 
     Node::<ipc::Service>::list(Config::global_config(), |node| {
