+++ iceoryx2-cli/iox2-service/src/command/details.rs
@@ -1,0 +1,45 @@
+// Copyright (c) 2025 Contributors to the Eclipse Foundation
+//
+// See the NOTICE file(s) distributed with this work for additional
+// information regarding copyright ownership.
+//
+// This program and the accompanying materials are made available under the
+// terms of the Apache Software License 2.0 which is available at
+// https://www.apache.org/licenses/LICENSE-2.0, or the MIT license
+// which is available at https://opensource.org/licenses/MIT.
+//
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+
+use anyhow::{Error, Result};
+use iceoryx2::prelude::*;
+use iceoryx2_cli::filter::Filter;
+use iceoryx2_cli::output::ServiceDescription;
+use iceoryx2_cli::Format;
+
+use crate::cli::OutputFilter;
+
+pub(crate) fn details(service_name: String, filter: OutputFilter, format: Format) -> Result<()> {
+    let mut error: Option<Error> = None;
+
+    ipc::Service::list(Config::global_config(), |service| {
+        if service_name == service.static_details.name().to_string() && filter.matches(&service) {
+            match format.as_string(&ServiceDescription::from(&service)) {
+                Ok(output) => {
+                    println!("{output}");
+                    CallbackProgression::Continue
+                }
+                Err(e) => {
+                    error = Some(e);
+                    CallbackProgression::Stop
+                }
+            }
+        } else {
+            CallbackProgression::Continue
+        }
+    })?;
+
+    if let Some(err) = error {
+        return Err(err);
+    }
+    Ok(())
+}
