+++ iceoryx2-cli/iox2-service/src/command/publish.rs
@@ -109,7 +109,7 @@ fn read_cli_msg_into_buffer(
     Ok(())
 }
 
-pub fn publish(options: PublishOptions, _format: Format) -> Result<()> {
+pub(crate) fn publish(options: PublishOptions, _format: Format) -> Result<()> {
     let node = NodeBuilder::new()
         .name(&NodeName::new(&options.node_name)?)
         .create::<ipc::Service>()?;
