+++ iceoryx2-cli/iox2-tunnel/src/command/mod.rs
@@ -1,4 +1,4 @@
-// Copyright (c) 2024 Contributors to the Eclipse Foundation
+// Copyright (c) 2025 Contributors to the Eclipse Foundation
 //
 // See the NOTICE file(s) distributed with this work for additional
 // information regarding copyright ownership.
@@ -10,11 +10,6 @@
 //
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-mod cli;
-mod format;
+mod zenoh;
 
-pub mod filter;
-pub mod output;
-
-pub use cli::*;
-pub use format::Format;
+pub(crate) use zenoh::*;
